Refactor delete command tests
- remove unused files
